+++ src/Services/Settlement/Settlement.Application/DTOs/MerchantPayoutData.cs
@@ -1,3 +1,3 @@
 ﻿namespace Settlement.Application.DTOs;
 
-public record MerchantPayoutData(Guid MerchantId, decimal GrossVolume, decimal Fees, string Currency);
+public record MerchantPayoutData(Guid MerchantId, long GrossVolume, long Fees, string Currency);
